Generic checking functions

These functions provide common, simple data checks.

chk_dummy(x)

chk_blank(x)

chk_equals(x, val)

chk_range(x, min, max, ...)

chk_pattern(x, pattern)

chk_max_length(x, len)

chk_text_miss(x, miss = getOption("testdat.miss_text"))

chk_text_nmiss(x, miss = getOption("testdat.miss_text"))

chk_unique(x)

chk_ascii(x)

chk_values(x, ..., miss = getOption("testdat.miss"))

chk_date_yyyymmdd(x)

chk_date_yyyymm(x)

chk_date_yyyy(x)

Arguments

x
A vector to check.

val
A scalar value for the equality check.

min
Minimum value for range check.

max
Maximum value for range check.

...
Vectors of valid values.

pattern
A str_detect() pattern to match.

len
Maximum string length.

miss
A vector of values to be treated as missing. The testdat.miss or testdat.miss_text option is used by default.

Value

A logical vector flagging records that have passed or failed the check.

See also

Examples

sales <- data.frame(
  sale_id = 1:5,
  date = c("20200101", "20200101", "20200102", "20200103", "2020003"),
  sale_price = c(10, 20, 30, 40, -1),
  book_title = c(
    "Phenomenology of Spirit",
    NA,
    "Critique of Practical Reason",
    "Spirit of Trust",
    "Empiricism and the Philosophy of Mind"
  ),
  stringsAsFactors = FALSE
)

lapply(sales, function(x) any(chk_blank(x))) # Any data missing?

#> $sale_id
#> [1] FALSE
#> 
#> $date
#> [1] FALSE
#> 
#> $sale_price
#> [1] FALSE
#> 
#> $book_title
#> [1] TRUE
#> 

with(sales, date[!chk_date_yyyymmdd(date)]) # Examine invalid dates

#> [1] "2020003"

all(chk_unique(sales$sale_id)) # Valid key?

#> [1] TRUE

# Clean out invalid prices
sales$sale_price[!chk_range(sales$sale_price, 0, Inf)] <- NA
sales

#>   sale_id     date sale_price                            book_title
#> 1       1 20200101         10               Phenomenology of Spirit
#> 2       2 20200101         20                                  <NA>
#> 3       3 20200102         30          Critique of Practical Reason
#> 4       4 20200103         40                       Spirit of Trust
#> 5       5  2020003         NA Empiricism and the Philosophy of Mind

# detect non-ASCII characters
chk_ascii(c("a", "\U1f642"))

#> [1]  TRUE FALSE