Deprecated expectation functions

These functions are deprecated.

See Generic Expectation Functions and Cross-dataset Expectations for current expectation functions.

expect_func(var, ...)

expect_join(data2, by = NULL, not = FALSE, flt = TRUE, data = get_testdata())

expect_similar(
  var,
  data2,
  var2,
  flt = TRUE,
  flt2 = flt,
  threshold = 0.05,
  min = 100,
  data = get_testdata()
)

Arguments

var
an unquoted variable name

...
arguments to pass to expect_allany()

data2
the dataset to compare against

by
a character vector of variables to join by. For details see dplyr join().

not
reverse the results of the check

flt
a filter specifying a subset of the dataset to test

data
data frame to test. The global test data is used by default.

var2
an unquoted variable name from data2

flt2
a filter specifying a subset of data2 to test

threshold
the maximum proportional difference allowed between the two categories

min
the minimum number of responses for a category to allow comparison. This avoidmall categories raising spurious errors

See also