Expectations: generic expectation helpers

These functions allow for testing of a dataset using an arbitrary function.

expect_allany(vars, func, flt = TRUE, data = get_testdata(),
  args = list(), allany = c(chk_filter_all, chk_filter_any),
  func_desc = NULL)

expect_all(...)

expect_any(...)

expect_func(var, ...)

Arguments

vars
a list of columns generated by vars()

func
a function that takes a vector as the first argument and returns a logical vector of the same length showing whether an element passed or failed

flt
a filter specifying a subset of the dataset to test

data
data frame to test. The global test data is used by default.

args
a named list of arguments to pass to func

allany
function to use to combine results for each vector

func_desc
A character function description to use in the expectation failure message.

...
arguments to pass to expect_allany()

var
an unquoted variable name

See also