Expectations: uniqueness checks

These functions test variables for uniqueness:

expect_unique tests a set of variables (vars) and fails if the variables collectively do not uniquely identify the observations

expect_unique_across tests a set of variables (vars) and fails if each observation does not have different values in each of the variables in vars.

expect_unique_combine tests a set of variables (vars) and fails if any values appear more than once across all of them.

expect_unique(
  vars,
  exclude = getOption("testdat.miss"),
  flt = TRUE,
  data = get_testdata()
)

expect_unique_across(
  vars,
  exclude = getOption("testdat.miss"),
  flt = TRUE,
  data = get_testdata()
)

expect_unique_combine(
  vars,
  exclude = getOption("testdat.miss"),
  flt = TRUE,
  data = get_testdata()
)

Arguments

vars
a list of columns generated by vars()

exclude
a vector of values to exclude from uniqueness check, defaults to getOption("testdat.miss") which is c(NA, "") by default. To include all values, set exclude=NULL.

flt
a filter specifying a subset of the dataset to test

data
data frame to test. The global test data is used by default.

Details

By default the uniqueness check excludes missing values (as specified by getOption("testdat.miss")). Setting exclude=NULL will include all values.

See also

Examples

student_fruit_preferences <- data.frame(
  student_id = c(1:5, NA, NA),
  apple = c(1, 1, 1, 1, 99, NA, NA),
  orange = c(2, 3, 2, 3, 99, NA, NA),
  banana = c(3, 2, 3, 2, 99, NA, NA),
  phone1 = c(123, 456, 789, 987, 654, NA, NA),
  phone2 = c(345, 678, 987, 567, 000, NA, NA)
)

# Check that key is unique, excluding NAs by default
expect_unique(vars(student_id), data = student_fruit_preferences)

# Check that key is unique, including NAs
try(expect_unique(vars(student_id), exclude = NULL, data = student_fruit_preferences))

#> Error : `student_fruit_preferences` has 2 duplicate records on variable `student_id`.
#> Filter: None

# Check each fruit has unique preference number
try(
expect_unique_across(
  vars(apple, orange, banana),
  data = student_fruit_preferences
)
)

#> Error : `student_fruit_preferences` has 1 records with duplicates across variables `apple, orange, banana`.
#> Filter: None

# Check each fruit has unique preference number, allowing multiple 99 (item skipped) codes
expect_unique_across(
  vars(apple, orange, banana),
  exclude = c(99, NA), data = student_fruit_preferences
)

# Check that each phone number appears at most once
try(expect_unique_combine(vars(phone1, phone2), data = student_fruit_preferences))

#> Error : `student_fruit_preferences` has 2 records with duplicate values across variables `phone1, phone2`.
#> Filter: None