Checking helper functions

These helper functions allowing easy checking over multiple columns and filtering to subsets of the data:

chk_filter applies func with args to var in data filtered with flt and returns the resulting logical vector.

chk_filter_vars applies func with args to vars in data filtered with flt and returns a dataframe containing the resulting logical vectors.

chk_filter_all and chk_filter_any are wrappers around chk_filter_vars which return a single logical vector produced by taking the conjunction and disjunction, respectively, of the columns in the output of chk_filter_vars

chk_filter_where works exactly the same as chk_filter_all except that the variables are specified not using dplyr::vars() (vars) but using bare tidy-select functions (where).

chk_filter(data, var, func, flt = TRUE, args = list())

chk_filter_vars(data, vars, func, flt = TRUE, args = list())

chk_filter_all(data, vars, func, flt = TRUE, args = list())

chk_filter_any(data, vars, func, flt = TRUE, args = list())

chk_filter_where(data, where, func, flt = TRUE, args = list())

Arguments

data
a tbl

var
a single column to check

func
a function to use use for checking. The provided function should take a single vector as input and return a logical vector.

flt
a logical expression to be applied as a filter for the check. This will be evaluated in the context of data

args
a list of additional arguments to be added to the function calls.

vars
a list of columns to check generated by vars()

where
<tidy-select> columns to check

Value

A logical vector flagging records that have passed or failed the check, with NA where records do not meet the filter

See also

Examples

# Check that every 4-cylinder car has an engine displacement of < 100 cubic
# inches
x <- chk_filter(mtcars, disp, chk_range, cyl == 4, list(min = 0, max = 100))
all(x, na.rm = TRUE)

#> [1] FALSE

# Check that every 4-cylinder car has an engine displacement of < 100 cubic
# inches AND < 100 horsepower
chk_filter_all(
  mtcars,
  c("disp", "hp"),
  chk_range,
  cyl == 4,
  list(min = 0, max = 100)
)

#>           Mazda RX4       Mazda RX4 Wag          Datsun 710      Hornet 4 Drive 
#>                  NA                  NA               FALSE                  NA 
#>   Hornet Sportabout             Valiant          Duster 360           Merc 240D 
#>                  NA                  NA                  NA               FALSE 
#>            Merc 230            Merc 280           Merc 280C          Merc 450SE 
#>               FALSE                  NA                  NA                  NA 
#>          Merc 450SL         Merc 450SLC  Cadillac Fleetwood Lincoln Continental 
#>                  NA                  NA                  NA                  NA 
#>   Chrysler Imperial            Fiat 128         Honda Civic      Toyota Corolla 
#>                  NA                TRUE                TRUE                TRUE 
#>       Toyota Corona    Dodge Challenger         AMC Javelin          Camaro Z28 
#>               FALSE                  NA                  NA                  NA 
#>    Pontiac Firebird           Fiat X1-9       Porsche 914-2        Lotus Europa 
#>                  NA                TRUE               FALSE               FALSE 
#>      Ford Pantera L        Ferrari Dino       Maserati Bora          Volvo 142E 
#>                  NA                  NA                  NA               FALSE 

# Check that columns made up of whole numbers are binary
chk_filter_where(
  mtcars,
  where(~ all(. %% 1 == 0)),
  chk_values,
  TRUE,
  list(0:1)
)

#>           Mazda RX4       Mazda RX4 Wag          Datsun 710      Hornet 4 Drive 
#>               FALSE               FALSE               FALSE               FALSE 
#>   Hornet Sportabout             Valiant          Duster 360           Merc 240D 
#>               FALSE               FALSE               FALSE               FALSE 
#>            Merc 230            Merc 280           Merc 280C          Merc 450SE 
#>               FALSE               FALSE               FALSE               FALSE 
#>          Merc 450SL         Merc 450SLC  Cadillac Fleetwood Lincoln Continental 
#>               FALSE               FALSE               FALSE               FALSE 
#>   Chrysler Imperial            Fiat 128         Honda Civic      Toyota Corolla 
#>               FALSE               FALSE               FALSE               FALSE 
#>       Toyota Corona    Dodge Challenger         AMC Javelin          Camaro Z28 
#>               FALSE               FALSE               FALSE               FALSE 
#>    Pontiac Firebird           Fiat X1-9       Porsche 914-2        Lotus Europa 
#>               FALSE               FALSE               FALSE               FALSE 
#>      Ford Pantera L        Ferrari Dino       Maserati Bora          Volvo 142E 
#>               FALSE               FALSE               FALSE               FALSE