Deprecated expectation functions

These functions are deprecated.

expect_func(var, ...)

expect_join(data2, by = NULL, not = FALSE, flt = TRUE, data = get_testdata())

expect_similar(
  var,
  data2,
  var2,
  flt = TRUE,
  flt2 = flt,
  threshold = 0.05,
  min = 100,
  data = get_testdata()
)

Arguments

var
An unquoted column name to test.

...
Arguments to pass to expect_allany().

data2
The data frame to compare against.

by
A character vector of columns to join by. See dplyr::join() for details.

not
Reverse the results of the check?

flt
<data-masking> A filter specifying a subset of the data frame to test.

data
A data frame to test. The global test data is used by default.

var2
An unquoted column name from data2.

flt2
A filter specifying a subset of data2 to test.

threshold
The maximum proportional difference allowed between the two categories.

min
The minimum number of responses for a category to allow comparison. This avoids small categories raising spurious errors.

See also